backend/requirements.txt
fastapi>=0.115.0
sqlmodel>=0.0.0
pyjwt[crypto]>=2.9.0
pydantic>=2.0.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
asyncpg>=0.29.0
email-validator>=2.0.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4
